cleanup fixtures and helpers
