+++ test/test_api.py
@@ -14,37 +14,16 @@
 from __future__ import print_function
 from __future__ import unicode_literals
 
-import itertools
-import os
+from helpers import default_query_classes, default_format_classes
 
 import pytest
-
-from rptk import RptkAPI
-
-
-@pytest.fixture(scope="session")
-def api():
-    """Construct an API instance with the test config."""
-    config_file = os.path.join(os.path.dirname(__file__), 'tests.conf')
-    return RptkAPI(config_file=config_file)
-
-
-@pytest.fixture(scope="session")
-def posix():
-    """Check whether we are on a POSIX system."""
-    return os.name == "posix"
-
-
-def classmap(api):
-    """Construct a cartesian product of the available classes."""
-    return itertools.product(api.query_class_loader.class_names,
-                             api.format_class_loader.class_names)
 
 
 class TestAPI(object):
     """Test cases for rptk python API."""
 
-    @pytest.mark.parametrize(("q", "f"), classmap(api()))
+    @pytest.mark.parametrize("q", default_query_classes().keys())
+    @pytest.mark.parametrize("f", default_format_classes().keys())
     def test_api(self, api, posix, q, f):
         """Test rptk python API."""
         if api.query_class_loader.get_class(name=q).posix_only and not posix:
